API

The list above reports all the functions implemented in SolarFunctions.jl. In the following, we group them into topic categories and we detail them.

SolarFunctions.aberration_correction

SolarFunctions.absolute_airmass

SolarFunctions.altitude2pressure

SolarFunctions.aod_bb_hulstrom1980

SolarFunctions.apparent_sidereal_greenwich_time

SolarFunctions.apparent_sun_longitude

SolarFunctions.ascending_moon_longitude

SolarFunctions.clearsky_haurwitz

SolarFunctions.clearsky_ineichen

SolarFunctions.clearsky_simplified_solis

SolarFunctions.day_angle

SolarFunctions.delta_T

SolarFunctions.dew_temperature2relative_humidity_oke2018

SolarFunctions.elliptic_obliquity

SolarFunctions.extraterrestrial_irradiance_asce

SolarFunctions.extraterrestrial_irradiance_nrel

SolarFunctions.extraterrestrial_irradiance_spencer1971

SolarFunctions.geocentric_latitude

SolarFunctions.geocentric_longitude

SolarFunctions.geocentric_sun_ascension

SolarFunctions.geocentric_sun_declination

SolarFunctions.heliocentric_latitude

SolarFunctions.heliocentric_longitude

SolarFunctions.heliocentric_radius

SolarFunctions.julian_century

SolarFunctions.julian_ephemeris_century

SolarFunctions.julian_ephemeris_day

SolarFunctions.julian_ephemeris_millenium

SolarFunctions.linke_turbidity_kasten1996

SolarFunctions.linke_turbidity_meteotest

SolarFunctions.mean_elliptic_obliquity

SolarFunctions.mean_moon_anomaly

SolarFunctions.mean_moon_elongation_from_sun

SolarFunctions.mean_sidereal_greenwich_time

SolarFunctions.mean_sun_anomaly

SolarFunctions.mean_sun_longitude

SolarFunctions.moon_latitude_argument

SolarFunctions.nutation_coefficients

SolarFunctions.nutation_longitude

SolarFunctions.nutation_obliquity

SolarFunctions.observer_local_hour

SolarFunctions.precipitable_water_gueymard94

SolarFunctions.pressure2altitude

SolarFunctions.radial_distance_equatorial_plane

SolarFunctions.radial_distance_rotational_axis

SolarFunctions.reduced_observer_latitude

SolarFunctions.relative_airmass_gueymard1993

SolarFunctions.relative_airmass_gueymard2003

SolarFunctions.relative_airmass_kasten1966

SolarFunctions.relative_airmass_pickering2002

SolarFunctions.relative_airmass_simple

SolarFunctions.relative_airmass_young1994

SolarFunctions.relative_airmass_youngirvine1967

SolarFunctions.relative_arimass_kastenyoung1989

SolarFunctions.relative_humidity2dew_temperature_oke2018

SolarFunctions.sun_ascension_parallax

SolarFunctions.sun_equatorial_horizontal_parallax

SolarFunctions.topocentric_apparent_elevation

SolarFunctions.topocentric_astronomical_azimuth

SolarFunctions.topocentric_azimuth

SolarFunctions.topocentric_elevation

SolarFunctions.topocentric_elevation_correction

SolarFunctions.topocentric_local_hour

SolarFunctions.topocentric_sun_ascension

SolarFunctions.topocentric_sun_declination

Solar position algorithm

The following functions implement the Solar Position Algorithm, as described in [1].

Julian period conversions

SolarFunctions.delta_T — Function

delta_T(datetime::DateTime)

Calculates the difference between Terrestrial Dynamical Time (TD) and Universal Time (UT)

It is evaluated as a piecewise polynomial, as per NASA, which in turn refer to [2, 3].

Note that $\Delta T$ is unknown for years before -1999 and after 3000. Values could be calculated for such intervals, although they are not intended to be used for these years.

SolarFunctions.julian_ephemeris_day — Function

julian_ephemeris_day(julian_day::Real, ΔT::Real)

A continuous count of days measured in uniform Ephemeris Time (or its successors), calculated as follows

\[J_{\text{ed}} = J_{\text{d}} + \frac{\Delta T}{86400}\]

where:

$J_{\text{d}}$ is the Julian day, which can be computed using Dates.datetime2julian

$\Delta T$ is the difference between Terrestrial Dynamical Time and Universal Time, which can be computed using delta_T.

SolarFunctions.julian_century — Function

julian_century(julian_day::Real)

A time interval of exactly 36,525 days (365.25 days × 100) used as a standard unit in astronomy, calculated as follows

\[J_{\text{c}} = \frac{J_{\text{d}} - 2451545}{36525}\]

where $J_{\text{d}}$ is the Julian day, which can be computed using Dates.datetime2julian..

SolarFunctions.julian_ephemeris_century — Function

julian_ephemeris_century(julian_ephemeris_day::Real)

A 36,525-day interval measured specifically in Ephemeris Time (or its modern dynamical time scales) for high-precision astronomical modeling. It is calculated as follows

\[J_{\text{ec}} = \frac{J_{\text{ed}} - 2451545}{36525}\]

where $J_{\text{ed}}$ is the julian_ephemeris_day.

SolarFunctions.julian_ephemeris_millenium — Function

julian_ephemeris_millenium(julian_ephemeris_century::Real)

A 1,000-year interval equal to 365,250 ephemeris days, defined within Ephemeris Time for long-term astronomical calculations. It is calculated as

\[J_{\text{em}} = \frac{J_{\text{ec}}}{10}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

Heliocentric quantites

SolarFunctions.heliocentric_longitude — Function

heliocentric_longitude(julian_ephemeris_millenium::Real)

A celestial object's angular distance north or south of the ecliptic plane as measured from the Sun.

It is calculated as a polynomial expression, where the coefficients are collected in $\mathcal{M} = \{\mathbf{M}_i\}_{i=1}^6$, with $\mathbf{M}_i \in \mathbb{R}^{n_i \times 3}$ being the coeffient matrix for the $i$-th polynomal coefficient $c_{\mu, i}$. The heliocentric longitude, $\mu_{\text{hel}}$, is thus calculated as

\[ \begin{align*} c_{\mu, i} &= \sum_{j = 1}^{n_i} M^{(j,1)}_i \cos(M^{(j,2)}_i + M^{(j,3)}_i J_{\text{em}}) \\ \mu_{\text{hel}} &= \frac{180}{\pi} \frac{1}{1e8} \sum_{i = 1}^6 c_{\ell, i} J_{\text{em}}^{i-1} \mod 360 \end{align*}\]

where $J_{\text{em}}$ is the julian_ephemeris_millenium. The $180/\pi$ term coverts $\mu_{\text{hel}}$ from radians to degrees. Moreover, the $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

The coefficient matrices in $\mathcal{M}$ are collected in SolarFunctions.EARTH_PERIODIC_TERMS.M.

SolarFunctions.heliocentric_latitude — Function

heliocentric_latitude(julian_ephemeris_millenium::Real)

A celestial object's angular position around the Sun measured in the ecliptic plane from a defined reference direction.

It is calculated as a polynomial expression, where the coefficients are collected in $\mathcal{L} = \{\mathbf{L}_i\}_{i=1}^2$, with $\mathbf{L}_i \in \mathbb{R}^{n_i \times 3}$ being the coeffient matrix for the $i$-th polynomal coefficient $c_{\lambda, i}$. The heliocentric latitude, $\lambda_{\text{hel}}$, is thus calculated as

\[ \begin{align*} c_{\lambda, i} &= \sum_{j = 1}^{n_i} L^{(j,1)}_i \cos(L^{(j,2)}_i + L^{(j,3)}_i J_{\text{em}}) \\ \lambda_{\text{hel}} &= \frac{180}{\pi} \frac{1}{1e8} \sum_{i = 1}^2 c_{b, i} J_{\text{em}}^{i-1} \end{align*}\]

where $J_{\text{em}}$ is the julian_ephemeris_millenium. The $180/\pi$ term coverts $\lambda_{\text{hel}}$ from radians to degrees.

The coefficient matrices in $\mathcal{L}$ are collected in SolarFunctions.EARTH_PERIODIC_TERMS.L.

SolarFunctions.heliocentric_radius — Function

heliocentric_radius(julian_ephemeris_millenium::Real)

The distance from the Sun to the celestial object in a chosen heliocentric coordinate system. In this case, it corresponds to the Earth-Sun distance.

It is calculated as a polynomial expression, where the coefficients are collected in $\mathcal{R} = \{\mathbf{R}_i\}_{i=1}^5$, with $\mathbf{R}_i \in \mathbb{R}^{n_i \times 3}$ being the coeffient matrix for the $i$-th polynomal coefficient $c_{R, i}$. The heliocentric radius, $R_{\text{hel}}$, is thus calculated as

\[ \begin{align*} c_{R, i} &= \sum_{j = 1}^{n_i} R^{(j,1)}_i \cos(R^{(j,2)}_i + R^{(j,3)}_i J_{\text{em}}) \\ R_{\text{hel}} &= \frac{180}{\pi} \frac{1}{1e8} \sum_{i = 1}^5 c_{r, i} J_{\text{em}}^{i-1} \end{align*}\]

where $J_{\text{em}}$ is the julian_ephemeris_millenium.

The coefficient matrices in $\mathcal{R}$ are collected in SolarFunctions.EARTH_PERIODIC_TERMS.R.

Geocentric quantities

SolarFunctions.geocentric_longitude — Function

geocentric_longitude(heliocentric_longitude::Real)

A celestial object's ecliptic longitude as measured from Earth’s center, referenced to the ecliptic plane and the vernal equinox. It is calculated as follows

\[\mu_{\text{geo}} = \mu_{\text{hel}} + 180 \mod 360\]

where $\mu_{\text{hel}}$ is the heliocentric_longitude [deg], and the $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

SolarFunctions.geocentric_latitude — Function

geocentric_latitude(heliocentric_latitude::Real)

A celestial object's angular distance north or south of the ecliptic plane as measured from Earth’s center. It is calculated as $\lambda_{\text{geo}} = -\lambda_{\text{hel}}$ where $\lambda_{\text{hel}}$ is the heliocentric_latitude [deg].

SolarFunctions.geocentric_sun_ascension — Function

geocentric_sun_ascension(
    apparent_sun_longitude::Real,
    elliptic_obliquity::Real,
    geocentric_latitude::Real
)

The angle measured eastward from the geocentric meridian, that locates a direction or point relative to Earth’s ellipsoid rather than its rotational axis. It is calculated as follows

\[ \sigma_{\text{geo}} = \text{atan2}(\sin\tilde{\mu}_{\odot} \cos\omega_{\text{ell}} - \tan\lambda_{\text{geo}}\sin\omega_{\text{ell}}, \cos\tilde{\mu}_{\odot})\]

where

$\tilde{\mu}_{\odot}$ corresponds to apparent_sun_longitude [deg]

$\omega_{\text{ell}}$ corresponds to elliptic_obliquity [deg]

$\lambda_{\text{geo}}$ corresponds to geocentric_latitude [deg]

SolarFunctions.geocentric_sun_declination — Function

geocentric_sun_declination(
    apparent_sun_longitude::Real,
    elliptic_obliquity::Real,
    geocentric_latitude::Real,
)

The Sun’s angular distance, north or south of Earth’s equatorial plane as measured from Earth’s center. It is calculated as follows

\[ \delta_{\text{geo}} = \text{asin}(\sin \lambda_{\text{geo}} \cos \omega_{\text{ell}} + \cos \lambda_{\text{geo}} \sin \omega_{\text{ell}} \sin \tilde{\mu}_{\odot}) \mod 360\]

where:

$\tilde{\mu}_{\odot}$ corresponds to apparent_sun_longitude [deg]

$\omega_{\text{ell}}$ corresponds to elliptic_obliquity [deg]

$\lambda_{\text{geo}}$ corresponds to geocentric_latitude [deg]

Moreover, the $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

Nutation quantities

SolarFunctions.mean_moon_elongation_from_sun — Function

mean_moon_elongation_from_sun(julian_ephemeris_century::Real)

The average angular separation between the Moon and the Sun as measured along the ecliptic. It is calculated as follows

\[ \bar{e}_{\bullet} = 297.85036 + 445267.111480J_{\text{ec}} − 0.0019142J_{\text{ec}}^2 + \frac{J_{\text{ec}}^3}{189474}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

SolarFunctions.mean_sun_anomaly — Function

mean_sun_anomaly(julian_ephemeris_century::Real)

The angular position of the Sun in its elliptical orbit, measured from perihelion and increasing uniformly in time. It is calculated as follows

\[ \bar{a}_{\odot} = 357.52772 + 35999.050340J_{\text{ec}} − 0.0001603J_{\text{ec}}^2 − \frac{J_{\text{ec}}^3}{300000}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

SolarFunctions.mean_moon_anomaly — Function

mean_moon_anomaly(julian_ephemeris_century::Real)

The uniformly increasing angular position of the Moon in its elliptical orbit measured from perigee. It is calculated as follows

\[ \bar{a}_{\bullet} = 134.96298 + 477198.867398J_{\text{ec}} + 0.0086972J_{\text{ec}}^2 + \frac{J_{\text{ec}}^3}{56250}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

SolarFunctions.moon_latitude_argument — Function

moon_latitude_argument(julian_ephemeris_century::Real)

The angle from the Moon’s ascending node to its position measured along its orbit, using mean (unperturbed) orbital elements. It is calculated as follows

\[ \lambda_{\bullet} = 93.27191 + 483202.017538J_{\text{ec}} - 0.0036825J_{\text{ec}}^2 + \frac{J_{\text{ec}}^3}{327270}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

SolarFunctions.ascending_moon_longitude — Function

ascending_moon_longitude(julian_ephemeris_century::Real)

The ecliptic longitude referenced to the mean equinox of the date, of the point where the Moon’s mean orbit crosses northward through the ecliptic. It is calculated as follows

\[ \mu_{\bullet} = 125.04452 - 1934.136261J_{\text{ec}} + 0.0020708J_{\text{ec}}^2 + \frac{J_{\text{ec}}^3}{450000}\]

where $J_{\text{ec}}$ is the julian_ephemeris_century.

SolarFunctions.nutation_coefficients — Function

nutation_coefficients(julian_ephemeris_century::Real)

Constructs the vector of weights for evaluating the nutation_longitude and nutation_obliquity. It is calculated as follows

\[ \begin{align*} w_i &= \sum_{j=1}^5 X_j Y_{ij} \\ \mathbf{X} &= [\bar{e}_{\bullet}, \bar{a}_{\odot}, \bar{a}_{\bullet}, \lambda_{\bullet}, \mu_{\bullet}]^\top \end{align*}\]

The coefficients in matrix $\mathbf{Y} \in \mathbb{R}^{m \times 5}$ are collected in SolarFunctions.EARTH_PERIODIC_TERMS.Y.

SolarFunctions.nutation_longitude — Function

nutation_longitude(julian_ephemeris_century::Real, nutation_coefficients::Vector)

The small periodic variation in Earth’s ecliptic longitude caused by gravitational torques from the Moon and Sun. It is calculated as follows

\[ \mu_{\text{nut}} = \frac{1}{3.6e7} \sum_{i=1}^n (\Psi_{i,1} + \Psi_{i,2} J_{\text{ec}}) \sin w_i\]

where:

$\boldsymbol{\Psi} \in \mathbb{R}^{m \times 2}$ is the coefficient matrix, whose terms are collected in SolarFunctions.EARTH_PERIODIC_TERMS.ψ

$J_{\text{ec}}$ is the julian_ephemeris_century.

$w_i$ is the $i$-th term of the nutation_coefficients vector

SolarFunctions.nutation_obliquity — Function

nutation_obliquity(julian_ephemeris_century::Real, nutation_coefficients::Vector)

The small periodic variation in Earth’s axial tilt (obliquity), $\omega_{\text{nut}}$, resulting from gravitational perturbations. It is calculated as follows

\[ \omega_{\text{nut}} = \frac{1}{3.6e7} \sum_{i=1}^n (E_{i,1} + E_{i,2} J_{\text{ec}}) \cos w_i\]

where:

$\boldsymbol{E} \in \mathbb{R}^{m \times 2}$ is the coefficient matrix, whose terms are collected in SolarFunctions.EARTH_PERIODIC_TERMS.ϵ

$J_{\text{ec}}$ is the julian_ephemeris_century.

$w_i$ is the $i$-th term of the nutation_coefficients vector

SolarFunctions.mean_elliptic_obliquity — Function

mean_elliptic_obliquity(julian_ephemeris_millenium::Real)

The angle between Earth’s equator and the mean (long-term averaged) ecliptic defined by an unperturbed elliptical Earth orbit. It is calculated as follows

\[ \begin{align*} U &= \frac{J_{\text{em}}}{10} \\ \bar{\omega}_{\text{ell}} &= 84381.448 − 4680.93U − 155.0U^2 + 1999.25U^3 − 51.38U^4 − \\ & \qquad 249.67U^5 − 39.05U^6 + 7.12U^7 + 27.87U^8 + 5.79U^9 + 2.45U^{10} \end{align*}\]

where $J_{\text{em}}$ is the julian_ephemeris_millenium.

SolarFunctions.elliptic_obliquity — Function

elliptic_obliquity(mean_elliptic_obliquity::Real, nutation_obliquity::Real)

The angle between Earth’s equator and the mean ecliptic defined as an ideal, unperturbed ellipse. It is calculated as follows

\[\omega_{\text{ell}} = \bar{\omega}_{\text{ell}} / 3600 + \omega_{\text{nut}}\]

where:

$\bar{\omega}_{\text{ell}}$ corresponds to mean_elliptic_obliquity [deg]

$\omega_{\text{nut}}$ corresponds to nutation_obliquity [deg]

Solar longitude

SolarFunctions.aberration_correction — Function

aberration_correction(heliocentric_radius::Real)

An adjustment applied to celestial coordinates to account for the apparent displacement caused by Earth’s motion through space. It is calculated as follows

\[\beta = -\frac{20.4898}{3600 R_{\text{hel}}}\]

where $R_{\text{hel}}$ corresponds to the heliocentric_radius [AU].

SolarFunctions.apparent_sun_longitude — Function

apparent_sun_longitude(
    geocentric_longitude::Real,
    nutation_longitude::Real,
    aberration_correction::Real
)

The Sun’s ecliptic longitude after applying corrections for nutation and aberration. It is calculated as follows

\[\tilde{\mu}_{\odot} = \mu_{\text{geo}} + \mu_{\text{nut}} + \beta\]

where:

$\mu_{\text{geo}}$ corresponds to geocentric_longitude [deg]

$\mu_{\text{nut}}$ corresponds to nutation_longitude [deg]

$\beta$ corresponds to aberration_correction [deg]

SolarFunctions.mean_sun_longitude — Function

mean_sun_longitude(julian_ephemeris_millenium::Real)

The Sun’s ecliptic longitude calculated from a uniformly moving fictitious Sun on the ecliptic. It is calculated as follows

\[ \bar{\mu}_{\odot} = 280.4664567 + 360007.6982779J_{\text{em}} + 0.03032028J_{\text{em}}^2 + \frac{J_{\text{em}}^3}{49931} - \frac{J_{\text{em}}^4}{15300} - \frac{J_{\text{em}}^5}{2000000}\]

where $J_{\text{em}}$ corresponds to the julian_ephemeris_millenium.

SolarFunctions.mean_sidereal_greenwich_time — Function

mean_sidereal_greenwich_time(julian_day::Real, julian_century::Real)

The hour angle of the mean vernal equinox at the Greenwich meridian reflecting Earth’s rotation relative to the fixed stars without nutation effects. It is calculated as follows

\[ \bar{t}_{\text{gw}} = 280.46061837 + 360.98564736629(J_{\text{d}} − 2451545.0) + 0.000387933J_{\text{c}}^2 - \frac{J_{\text{c}}^3}{38710000} \mod 360\]

where $J_{\text{d}}$ is the Julian Day, which can be computed using Dates.datetime2julian. Moreover, the $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

SolarFunctions.apparent_sidereal_greenwich_time — Function

apparent_sidereal_greenwich_time(
    mean_sidereal_greenwich_time::Real,
    nutation_longitude::Real, 
    elliptic_obliquity::Real
)

Greenwich sidereal time corrected for the effects of nutation giving the true rotation angle relative to the apparent equinox. It is calculated as

\[\tilde{t}_{\text{gw}} = \bar{t}_{\text{gw}} + \mu_{\text{nut}} \cos \omega_{\text{ell}}\]

where:

$\bar{t}_{\text{gw}}$ corresponds to mean_sidereal_greenwich_time

$\mu_{\text{nut}}$corresponds to nutation_longitude

$\omega_{\text{ell}}$corresponds to elliptic_obliquity

SolarFunctions.observer_local_hour — Function

observer_local_hour(
    observer_longitude::Real, 
    apparent_sidereal_greenwich_time::Real, 
    geocentric_sun_ascension::Real
)

The angle between the observer’s local meridian and a celestial object measured westward on the celestial sphere. It is calculated as

\[h_{\text{obs}} = \tilde{t}_{\text{gw}} + \mu_{\text{obs}} - \sigma_{\text{geo}} \mod 360\]

where:

$\mu_{\text{obs}}$ corresponds to the observer_longitude

$\tilde{t}_{\text{gw}}$ corresponds to apparent_sidereal_greenwich_time

$\sigma_{\text{geo}}$ corresponds to geocentric_sun_ascension

Moreover, the $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

Solar parallax

SolarFunctions.reduced_observer_latitude — Function

reduced_observer_latitude(observer_latitude::Real)

The angle whose tangent equals the tangent of the geodetic latitude scaled by Earth’s polar-to-equatorial radius ratio, representing the point’s projection onto the surface of the reference ellipsoid. It is calculated as follows

\[\lambda_{\text{red}} = \text{atan}(0.99664719\tan(\lambda_{\text{obs}}))\]

where $\lambda_{\text{obs}}$ corresponds to the observer_latitude.

SolarFunctions.radial_distance_equatorial_plane — Function

radial_distance_equatorial_plane(
    observer_latitude::Real,
    observer_altitude::Real,
    reduced_observer_latitude::Real
)

The perpendicular distance of a point from Earth’s equatorial plane. It is calculated as follows

\[d_{\text{eqt}} = \cos \lambda_{\text{red}} + A_{\text{obs}} \frac{\cos \lambda_{\text{obs}}}{6378140}\]

where:

$\lambda_{\text{obs}}$ corresponds to observer_latitude

$A_{\text{obs}}$ corresponds to observer_altitude

$\lambda_{\text{red}}$ corresponds to reduced_observer_latitude

SolarFunctions.radial_distance_rotational_axis — Function

radial_distance_rotational_axis(
    observer_latitude::Real, 
    observer_altitude::Real,
    reduced_observer_latitude::Real
)

The perpendicular distance of a point from Earth’s spin axis. It calculates as

\[d_{\text{rot}} = 0.99664719\sin \lambda_{\text{red}} + A_{\text{obs}} \frac{\sin \lambda_{\text{obs}}}{6378140}\]

where:

$\lambda_{\text{obs}}$ corresponds to observer_latitude

$A_{\text{obs}}$ corresponds to observer_altitude

$\lambda_{\text{red}}$ corresponds to reduced_observer_latitude

SolarFunctions.sun_equatorial_horizontal_parallax — Function

sun_equatorial_horizontal_parallax(heliocentric_radius::Real)

The angle between the Sun’s direction as seen from Earth’s center and from a point on the equator at sea level when the Sun is on the horizon. It is calculated as follows

\[\xi_{\text{hor}} = \frac{8.794}{3600 R_{\text{hel}}}\]

where $R_{\text{hel}}$ corresponds to the heliocentric_radius, in astronomical units.

SolarFunctions.sun_ascension_parallax — Function

sun_ascension_parallax(
    radial_distance_equatorial_plane::Real, 
    sun_equatorial_horizontal_parallax::Real, 
    geocentric_sun_declination::Real, 
    observer_local_hour::Real
)

The correction to the Sun’s right ascension due to the difference between geocentric and topocentric perspectives. It is calculated as follows

\[ \xi_{\text{asc}} = \text{atan2}(-d_{\text{eqt}} \sin \xi_{\text{hor}} \sin h_{\text{obs}}, \cos \delta_{\text{geo}} - d_{\text{eqt}} \sin \xi_{\text{hor}} \cos h_{\text{obs}})\]

where:

$d_{\text{eqt}}$ corresponds to radial_distance_equatorial_plane) [m]

$\xi_{\text{hor}}$ corresponds to sun_equatorial_horizontal_parallax [deg]

$\delta_{\text{geo}}$ corresponds to geocentric_sun_declination [deg]

$h_{\text{obs}}$ corresponds to observer_local_hour [deg]

Topocentric quantities

SolarFunctions.topocentric_sun_ascension — Function

topocentric_sun_ascension(
    geocentric_sun_ascension::Real, 
    sun_ascension_parallax::Real
)

The Sun’s right ascension as viewed from the observer’s actual location on Earth’s surface. It is calculated as follows

\[\sigma_{\text{top}} = \sigma_{\text{geo}} + \xi_{\text{asc}}\]

where:

$\sigma_{\text{geo}}$ corresponds to geocentric_sun_ascension [deg]

$\xi_{\text{asc}}$ corresponds to sun_ascension_parallax [deg]

SolarFunctions.topocentric_local_hour — Function

topocentric_local_hour(
    observer_local_hour::Real,
    sun_ascension_parallax::Real
)

The hour angle of a celestial object as seen from the observer’s exact location on Earth’s surface rather than its center. It is calculated follows

\[h_{\text{top}} = h_{\text{obs}} - \xi_{\text{asc}}\]

where:

$h_{\text{obs}}$ corresponds to the observer_local_hour [deg]

$\xi_{\text{asc}}$ corresponds to the sun_ascension_parallax [deg]

SolarFunctions.topocentric_sun_declination — Function

topocentric_sun_declination(
    radial_distance_equatorial_plane::Real, 
    radial_distance_rotational_axis::Real, 
    sun_equatorial_horizontal_parallax::Real, 
    geocentric_sun_declination::Real, 
    observer_local_hour::Real, 
    sun_ascension_parallax::Real
)

The Sun’s declination as viewed from the observer’s actual location on Earth’s surface. It is calculated as follows:

\[ \delta_{\text{top}} = \text{atan2}((\sin \delta_{\text{geo}} - d_{\text{rot}}\sin \xi_{\text{hor}})\cos \xi_{\text{asc}}, \cos \delta_{\text{geo}} - d_{\text{eqt}}\sin \xi_{\text{hor}} \cos h_{\text{obs}})\]

where:

$d_{\text{eqt}}$ corresponds to radial_distance_equatorial_plane [m]

$d_{\text{rot}}$ corresponds to radial_distance_rotational_axis [m]

$\xi_{\text{hor}}$ corresponds to sun_equatorial_horizontal_parallax [deg]

$\delta_{\text{geo}}$ corresponds to geocentric_sun_declination [deg]

$h_{\text{obs}}$ corresponds to observer_local_hour [deg]

$\xi_{\text{asc}}$ corresponds to sun_ascension_parallax [deg]

SolarFunctions.topocentric_apparent_elevation — Function

topocentric_apparent_elevation(
    observer_latitude::Real,
    topocentric_sun_declination::Real,
    topocentric_local_hour::Real
)

Topocentric elevation angle without atmospheric refraction correction. It is calculated as follows

\[ \tilde{\eta}_{\text{top}} = \text{asin}(\sin \lambda_{\text{obs}} \sin \delta_{\text{top}} + \cos \lambda_{\text{obs}} \cos \delta_{\text{top}} \cos h_{\text{top}})\]

where:

$\lambda_{\text{obs}}$ corresponds to observer_latitude [deg]

$\delta_{\text{top}}$ corresponds to topocentric_sun_declination [deg]

$h_{\text{top}}$ corresponds to topocentric_local_hour [deg]

SolarFunctions.topocentric_elevation_correction — Function

topocentric_elevation_correction(
    temperature::Real,
    pressure::Real,
    topocentric_apparent_elevation::Real
)

The adjustment applied to an object’s observed altitude to account for the bending of its light as it passes through Earth’s atmosphere. It is calculated as follows:

\[ \hat{\eta}_{\text{top}} = \frac{P}{1010} \frac{283}{273 + T} \frac{1.02}{60\tan(\tilde{\eta}_{\text{top}} + \frac{10.3}{\tilde{\eta}_{\text{top}} + 5.11})}\]

where:

$T$ corresponds to the local temperature [C]

$P$ corresponds to the local pressure [mbar]

$\tilde{\eta}_{\text{top}}$ corresponds to the topocentric_apparent_elevation [deg]

SolarFunctions.topocentric_elevation — Function

function topocentric_elevation(
    topocentric_apparent_elevation::Real, 
    topocentric_elevation_correction::Real
)

The angle between the Sun and the observer’s local horizon, measured at the observer’s location. It is calculated as follows

\[\eta_{\text{top}} = \tilde{\eta}_{\text{top}} + \hat{\eta}_{\text{top}}\]

where:

$\hat{\eta}_{\text{top}}$ corresponds to the topocentric_elevation_correction [deg]

$\tilde{\eta}_{\text{top}}$ corresponds to the topocentric_apparent_elevation [deg]

SolarFunctions.topocentric_astronomical_azimuth — Function

topocentric_astronomical_azimuth(
    observer_latitude::Real, 
    topocentric_sun_declination::Real, 
    topocentric_local_hour::Real
)

The Sun’s azimuth measured from true north eastward as seen from the observer’s location, based on astronomical (not navigational) convention. It is calculated as follows

\[ \alpha_{\star, \text{top}} = \text{atan2}(\sin h_{\text{top}}, \cos h_{\text{obs}} \sin \lambda_{\text{obs}} - \tan \delta_{\text{top}} \cos \lambda_{\text{obs}}) \mod 360\]

where:

$\lambda_{\text{obs}}$ corresponds to observer_latitude [deg]

$\delta_{\text{top}}$ corresponds to topocentric_sun_declination [deg]

$h_{\text{top}}$ corresponds to topocentric_local_hour [deg]

Moreover, the $\mod(\cdot)$ operator ensures that the results remains in the [0, 360] range.

SolarFunctions.topocentric_azimuth — Function

topocentric_azimuth(topocentric_astronomical_azimuth::Real)

The Sun’s azimuth from the observer’s location relative to a defined reference direction, typically true north, on the horizon. It is calculated as follows

\[\alpha_{\text{top}} = \alpha_{\star, \text{top}} + 180 \mod 360\]

where $\alpha_{\star, \text{top}}$ is the topocentric_astronomical_azimuth [deg].

Clear-sky

Implements different models to calculate the GHI, DNI, and DHI components under the clear sky assumption.

SolarFunctions.clearsky_ineichen — Function

clearsky_ineichen(
    sun_apparent_elevation::Real,
    observer_altitude::Real,
    absolute_airmass::Real,
    linke_turbidity::Real,
    extraterrestial_radiation::Real,
    perez_enhancement::Bool
)

Implements the Ineichen and Perez clear sky model for global horizontal irradiance (GHI), direct normal irradiance (DNI), and calculates the clear-sky diffuse horizontal (DHI) component as the difference between GHI and DNI*cos(zenith) as presented in [4, 5]. A report on clear sky models found the Ineichen/Perez model to have excellent performance with a minimal input data set [6]. Default values for monthly Linke turbidity provided by SoDa [7, 8].

Arguments

sun_apparent_elevation::Real - [deg] Refraction corrected solar elevation angle

observer_altitude::Real - [m] Altitude above sea level.

absolute_airmass::Real - Pressure corrected airmass.

linke_turbidity::Real - Linke Turbidity.

extraterrestial_radiation::Real - [W/m^2] Extraterrestrial irradiance.

perez_enhancement::Bool - Controls if the Perez enhancement factor should be applied. Setting to true may produce spurious results for times when the Sun is near the horizon and the airmass is high.

Returns

dni::Real - [W/m^2] direct normal irradiance

dhi::Real - [W/m^2] direct horizontal irradiance

ghi::Real - [W/m^2] global horizoantal irradiance

SolarFunctions.clearsky_haurwitz — Function

clearsky_haurwitz(sun_apparent_elevation::Real)

Implements the Haurwitz clear sky model for global horizontal irradiance (GHI) as presented in [9, 10]. A report on clear sky models found the Haurwitz model to have the best performance in terms of average monthly error among models which require only the Sun's elevation [6].

Arguments

sun_apparent_elevation::Real - [deg] The apparent (refraction corrected) Sun's elevation angle in degrees.

Returns

ghi::Real - [W/m^2] The modeled global horizonal irradiance in provided by the Haurwitz clear-sky model.

SolarFunctions.clearsky_simplified_solis — Function

clearsky_simplified_solis(
    sun_apparent_elevation::Real,
    aod700::Real,
    precipitable_water::Real,
    pressure::Real,
    extraterrestial_radiation::Real
)

Calculate the clear sky GHI, DNI, and DHI according to the simplified Solis model. Reference [11] describes the accuracy of the model as being 15, 20, and 18 W/m^2 for the beam, global, and diffuse components, respectively. Reference [12] provides comparisons with other clear sky models.

Arguments

sun_apparent_elevation::Real - [degrees] The apparent elevation of the sun above the horizon.

aod700::Real - The aerosol optical depth at 700 nm. Algorithm derived for values between 0 and 0.45.

precipitable_water::Real - [cm] The precipitable water of the atmosphere. Algorithm derived for values between 0.2 and 10 cm. Values less than 0.2 will be assumed to be equal to 0.2.

pressure::Real - [Pa] The atmospheric pressure. Algorithm derived for altitudes between sea level and 7000 m, or 101325 and 41000 Pascals.

extraterrestial_radiation::Real - [W/m^2] Extraterrestrial irradiance. The units of extraterrestial_radiation determine the units of the output.

Returns

dni::Real - [W/m^2] direct normal irradiance

dhi::Real - [W/m^2] direct horizontal irradiance

ghi::Real - [W/m^2] global horizoantal irradiance

Atmosphere

Linke turbidity

SolarFunctions.aod_bb_hulstrom1980 — Function

aod_bb_hulstrom1980(aod380::Real, aod500::Real)

Approximate broadband aerosol optical depth (AOD) as proposed in [13, 14], which develops a correlation for broadband aerosol optical depth using two wavelengths, 380 nm and 500 nm. Arguments aod380 and aod500 are the AOD measured at 380 nm and 500 nm, respectively.

SolarFunctions.linke_turbidity_meteotest — Function

linke_turbidity_meteotest(
    observer_latitude::Real,
    observer_longitude::Real,
    month_index::Int,
)

Bilinear interpolation of the Linke turbidity array, for the corrsponding latitude, longitude, and month index values.

In the the Linke turbidity array, rows represent global latitudes from 90 to -90 degrees; columns represent global longitudes from -180 to 180; depth (third dimension) represents months of the year from January (1) to December (12).

The array is based on the tabular data in [15, Sec. 8].

SolarFunctions.linke_turbidity_kasten1996 — Function

linke_turbidity_kasten1996(
    absolute_airmass::Real, 
    precipitable_water::Real, 
    aod_bb::Real
)

Calculate Linke turbidity using Kasten pyrheliometric formula.

Note that broadband aerosol optical depth (AOD) can be approximated by AOD measured at 700 nm according to Molineaux [16]. Bird and Hulstrom offer an alternate approximation using AOD measured at 380 nm and 500 nm.

Based on original implementation by Armel Oumbe. Note that these calculations are only valid for airmass less than 5 and precipitable water less than 5 cm.

Arguments

absolute_airmass::Real - Pressure-adjusted airmass.

precipitable_water::Real - [cm] Precipitable water.

aod_bb::Real - broadband AOD.

Altitude and pressure

SolarFunctions.pressure2altitude — Function

pressure2altitude(pressure::Real)

Determine altitude, $A$ [m], from site pressure $P$ [Pa]. The following assumptions [17] hold:

Base pressure: 101325 Pa

Temperature at zero altitude: 288.15 K

Gravitational acceleration: 9.80665 m/s^2

Lapse rate: -6.5E-3 K/m

Gas constant for air: 287.053 J/(kg K)

Relative Humidity: 0%

\[ A = 44331.5 - 4946.62P^{0.190263}\]

SolarFunctions.altitude2pressure — Function

altitude2pressure(altitude::Real)

Determine site pressure, $P$ [Pa], from altitude $A$ [m]. It follows the same assumption of pressure2altitude.

\[ P = 100.0\left(\frac{44331.514 - A}{11880.516}\right)^{\frac{1}{0.1902632}}\]

Airmass

SolarFunctions.absolute_airmass — Function

absolute_airmass(relative_airmass::Real, pressure::Real=101325.0)

Determine absolute (pressure-adjusted) airmass, $a_\text{abs}$ from relative airmass, $a_\text{reò}$, and pressure, $P$ [Pa], [18].

\[ a_{\text{abs}} = a_{\text{rel}} \frac{P}{P_{\text{atm}}} \]

where $P_{\text{atm}}$ is the atmospheric pressure.

The following functions return the relative (not pressure-adjusted) airmass at sea level, for different models. Some models use apparent (refraction-adjusted) zenith angle while other models use true (not refraction-adjusted) zenith angle. Apparent zenith angles should be calculated at sea level. A comparison among several models is reported in [6].

SolarFunctions.relative_airmass_simple — Function

relative_airmass_simple(sun_apparent_elevation::Real)

Calculates the relative airmass as $m_{\text{air, rel}} = -\sec \tilde{\eta}_{\text{top}}$, where \tilde{\eta}{\text{top}} is the Sun's [topocentricapparentelevation](@ref) [deg]. Note that this returns -Inf at `sunapparent_elevation=0`.

SolarFunctions.relative_airmass_kasten1966 — Function

relative_airmass_kasten1966(sun_apparent_elevation::Real)

Computes the relative airmass as proposed in [19] as follows

\[m_{\text{air, rel}} = \frac{1}{-\cos \tilde{\eta}_{\text{top}} + 0.15((3.885 + \tilde{\eta}_{\text{top}})^(-1.253))}\]

where $\tilde{\eta}_{\text{top}}$ is the Sun's topocentric_apparent_elevation [deg].

SolarFunctions.relative_airmass_youngirvine1967 — Function

relative_airmass_youngirvine1967(sun_elevation::Real)

Computes the relative airmass as proposed in [20] as follows

\[ m_{\text{air, rel}} = -\sec \eta_{\text{top}} (1 - 0.0012(\sec \eta_{\text{top}}^2 + 1))\]

where $\eta_{\text{top}}$ is the Sun's topocentric_elevation [deg].

SolarFunctions.relative_arimass_kastenyoung1989 — Function

relative_arimass_kastenyoung1989(sun_apparent_elevation::Real)

Computes the relative airmass as proposed in [21] as follows

\[ m_{\text{air, rel}} = \frac{1}{-\cos \tilde{\eta}_{\text{top}} + 0.50572((6.07995 + \tilde{\eta}_{\text{top}})^{-1.6364})}\]

where $\tilde{\eta}_{\text{top}}$ is the Sun's topocentric_apparent_elevation [deg].

SolarFunctions.relative_airmass_gueymard1993 — Function

relative_airmass_gueymard1993(sun_apparent_elevation::Real)

Computes the relative airmass as proposed in [18, 22] as follows

\[ m_{\text{air, rel}} = \frac{1}{-\cos \tilde{\eta}_{\text{top}} + 0.00176759(90 - \tilde{\eta}_{\text{top}})((4.37515 + \tilde{\eta}_{\text{top}})^{- 1.21563})}\]

where $\tilde{\eta}_{\text{top}}$ is the Sun's topocentric_apparent_elevation [deg].

SolarFunctions.relative_airmass_young1994 — Function

relative_airmass_young1994(sun_elevation::Real)

Computes the relative airmass as proposed in [23] as follows

\[ m_{\text{air, rel}} = \frac{1.002432\cos \eta_{\text{top}}^2 + 0.148386\cos \eta_{\text{top}} + 0.0096467}{\cos \eta_{\text{top}}^3 + 0.149864\cos \eta_{\text{top}}^2 + 0.0102963\cos m_{\text{air, rel}} + 0.000303978}\]

where $\eta_{\text{top}}$ is the Sun's topocentric_elevation [deg].

SolarFunctions.relative_airmass_pickering2002 — Function

relative_airmass_pickering2002(sun_apparent_elevation::Real)

Computes the relative airmass as proposed in [24] as follows

\[ m_{\text{air, rel}} = \frac{1}{\sin(\tilde{\eta}_{\text{top}} + 244.0 / (165 + 47.0\tilde{\eta}_{\text{top}}^{1.1}))}\]

where $\eta_{\text{top}}$ is the Sun's topocentric_elevation [deg].

SolarFunctions.relative_airmass_gueymard2003 — Function

relative_airmass_gueymard2003(sun_apparent_elevation::Real)

Computes the relative airmass as proposed in [25, 26] as follow as follows

\[ m_{\text{air, rel}} = -\cos \tilde{\eta}_{\text{top}} + \frac{0.48353(90 - \tilde{\eta}_{\text{top}})^{0.095846}}{(6.741 + \tilde{\eta}_{\text{top}})^{1.754}}\]

where $\eta_{\text{top}}$ is the Sun's topocentric_elevation [deg].

Precipitable water and dew temperature

SolarFunctions.precipitable_water_gueymard94 — Function

precipitable_water_gueymard94(
    air_temperature::Real, 
    relative_humidity::Real
)

Calculates precipitable water (cm) from ambient air temperature [C] and relatively humidity [%] using an empirical model. The accuracy of this method is approximately 20% for moderate PW (1-3 cm) and less accurate otherwise.

The model was developed by expanding Eq. 1 in [27, Eq. 1] and using [27, Eq. 2]:

\[ \begin{align*} Pw = 0.1 H_v \rho_v \\ \rho_v = 216.7 R_H e_s \end{align*}\]

$Pw$ is the precipitable water (cm), $H_v$ is the apparent water vapor scale height [km] and :math:$\rho_v$ is the surface water vapor density [g/m^3]. The expression for $H_v$ results from [27, Eq. 4]:

\[ H_v = 0.4976 + 1.5265 \frac{T}{273.15} + \exp \left(13.6897 \frac{T}{273.15} - 14.9188 \left( \frac{T}{273.15} \right)^3 \right)\]

In the expression for $\rho_v$, $e_s$ is the saturation water vapor pressure [millibar]. The expression for $e_s$ results from [28, Eq. 1]

\[ e_s = \exp \left(22.330 - 49.140 \frac{100}{T} - 10.922 \left(\frac{100}{T}\right)^2 - 0.39015 \frac{T}{100} \right)\]

SolarFunctions.dew_temperature2relative_humidity_oke2018 — Function

dew_temperature2relative_humidity_oke2018(
    air_temperature::Real,
    dew_temperature::Real
)

Calculate relative humidity [%] from dewpoint temperature [°C] using the Magnus equation. Air temperature is expressed in °C. The Magnus equation coefficients (A, B, C) default to the values recommended by [29].

SolarFunctions.relative_humidity2dew_temperature_oke2018 — Function

relative_humidity2dew_temperature_oke2018(
    air_temperature::Real,
    relative_humidity::Real
)

Calculate dewpoint temperature using the Magnus equation. It corresponds to the inverse of dew_temperature2relative_humidity_oke2018.

Irradiance

SolarFunctions.day_angle — Function

day_angle(datetime::DateTime, offset::Int)

Calculates the day angle [deg] for the Earth's orbit around the Sun. For the Spencer method, offset=1; for the ASCE method, offset=0.

SolarFunctions.extraterrestrial_irradiance_spencer1971 — Function

extraterrestrial_irradiance_spencer1971(
    day_angle::Real, 
    solar_constant::Real
)

Calculates the extraterrestrial irradiance [W/m^2], as proposed in [30], as follows,

\[ R_{\text{ext}} = \Sigma(1.00011 + 0.034221\cos H + 0.00128\sin H + 0.000719\cos 2H + 7.7e-05\sin 2H)\]

where

$H$ corresponds to day_angle [deg]

$\Sigma$ corresponds to solar_constant [W/m^2]

SolarFunctions.extraterrestrial_irradiance_asce — Function

extraterrestrial_irradiance_asce(
    day_angle::Real,
    solar_constant::Real
)

Calculates the extraterrestrial irradiance using the ASCE evapotranspiration equation [31] as follows

\[ R_{\text{ext}} = \Sigma(1 + 0.033\cos(H))\]

where

$H$ corresponds to day_angle [deg]

$\Sigma$ corresponds to solar_constant [W/m^2]

SolarFunctions.extraterrestrial_irradiance_nrel — Function

extraterrestrial_irradiance_nrel(
    heliocentric_radius::Real,
    solar_constant::Real
)

Calculates the extraterrestrial irradiance using the heliocentric radius [AU], i.e., the Earth-Sun distance, as implemented by the NREL SPA [1], as follows

\[ R_{\text{ext}} = \Sigma / R_{\text{hel}}^2\]

where

$R_{\text{hel}}$ corresponds to heliocentric_radius: [AU]

$\Sigma$ corresponds to solar_constant: [W/m^2]